Several tweaks

+++ ngsF-HMM.hpp
@@ -56,7 +56,7 @@ typedef struct {
   double*** a;
   double* freq;
   double*** e;
-  uint64_t** path;
+  uint** path;
   double*** marg_prob;
   double* indF;
   double* lkl;
